feat: set database name to workbookId on creation & reorganize code
