+++ plugins/foreign-db-extractor/package.json
@@ -28,11 +28,11 @@
   "type": "module",
   "license": "ISC",
   "dependencies": {
-    "@aws-sdk/client-rds": "^3.458.0",
-    "@aws-sdk/client-s3": "^3.458.0",
+    "@aws-sdk/client-rds": "^3.496.0",
+    "@aws-sdk/client-s3": "^3.496.0",
     "@flatfile/api": "^1.6.3",
     "@flatfile/listener": "^0.4.0",
-    "@flatfile/util-file-buffer": "^0.1.3",
+    "@flatfile/util-file-buffer": "^0.2.0",
     "mssql": "^10.0.1"
   }
 }
